refactor(structure): NavContext→NavProvider, unify logos on *Mark
- Rename NavContext.tsx → NavProvider.tsx to match its NavProvider/useNav
  exports and the sibling *Provider files in context/.
- Unify the brand mark components on the *Mark suffix (GitHub's official
  term for an icon-only mark): AnthropicLogo→AnthropicMark,
  GeminiLogo→GeminiMark; GitHubMark unchanged.
- usePwa already relocated to hooks/ on main; no action here.
